requirements.txt
halftone==0.0.1
pillow==11.0.0
svgwrite==1.4.3